+++ tests/Fixtures/ArticleQueryFilter.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Teoprayoga\Teorion\Tests\Fixtures;
+
+use Teoprayoga\Teorion\QueryFilter;
+
+class ArticleQueryFilter extends QueryFilter
+{
+    public function filters(): array
+    {
+        return [];
+    }
+}